Lazy load custom resources from the load path
When a resource with a type starting with "Regentanz::Resources::" is encountered it is dynamically created. Before, this required that the class was loaded, but this will convert the resource type/class reference into a path, using the standard Ruby convention for the file names of classes, and then `require` that path, and attempt to create the resource again. Failure to load the file or if the file does not include any class with the expected name will result in the same error as before.