Scan for a config file and use it to populate the load path
In order for the autoloading of custom resources to work there needs to be a way to tell Regentanz where the resources are. This adds a quick and dirty mechanism where we scan the current directory, and recursively each parent directory for a `.regentanz.{yaml,yml,json}`, and if we find one, and it has a `load_paths` list, we add those paths to the load path.