Add support for CloudFormation YAML shorthands
In the standard support for YAML, the intrinsic functions can be written using YAML tags instead of using the traditional map-based approach.

Given that the shorthand is limited to the YAML format, it doesn't seem to add much value, as it only saves a few keystrokes per invocation, but it would make sense to be able to bring your existing YAML files into the tool.

For that use case, I guess one could argue that it would make more sense to output the template in YAML format as well.